Fix clippy warnings and formatting for CI
- Fix inconsistent digit grouping in currency amounts (use 1_000_000 instead of 10_000_00)
- Replace manual div_ceil implementations with .div_ceil()
- Fix needless_range_loop by using iterators with enumerate()
- Fix await_holding_lock in ringkernel-cpu kernel
- Fix redundant closures and unnecessary borrows
- Fix field_reassign_with_default warnings
- Remove unused imports and variables
- Fix identity_op and erasing_op warnings
- Replace for_kv_map with .keys() iteration
- Fix duplicate cfg attribute in simd module
- Use copy_from_slice instead of manual memcpy loops

+++ crates/ringkernel/examples/ping_pong.rs
@@ -32,7 +32,9 @@ async fn main() -> Result<()> {
     println!("Runtime created with K2K messaging enabled\n");
 
     // Get the K2K broker
-    let broker = runtime.k2k_broker().expect("K2K broker should be available");
+    let broker = runtime
+        .k2k_broker()
+        .expect("K2K broker should be available");
     println!("K2K broker obtained");
 
     // Register ping and pong endpoints
@@ -58,14 +60,12 @@ async fn main() -> Result<()> {
         // PING sends to PONG
         let timestamp = HlcTimestamp::now(round);
         let envelope = MessageEnvelope::empty(
-            round as u64,      // source_id (simulated)
-            round as u64 + 1,  // dest_id (simulated)
+            round,     // source_id (simulated)
+            round + 1, // dest_id (simulated)
             timestamp,
         );
 
-        let receipt = ping_endpoint
-            .send(pong_id.clone(), envelope)
-            .await?;
+        let receipt = ping_endpoint.send(pong_id.clone(), envelope).await?;
 
         match receipt.status {
             DeliveryStatus::Delivered => {
@@ -91,9 +91,7 @@ async fn main() -> Result<()> {
                 response_timestamp,
             );
 
-            let _response_receipt = pong_endpoint
-                .send(ping_id.clone(), response)
-                .await?;
+            let _response_receipt = pong_endpoint.send(ping_id.clone(), response).await?;
 
             println!("         PONG -> PING [responded]\n");
         }
@@ -111,10 +109,7 @@ async fn main() -> Result<()> {
             .send_priority(pong_id.clone(), envelope, priority)
             .await?;
 
-        println!(
-            "Sent priority {} message: {:?}",
-            priority, receipt.status
-        );
+        println!("Sent priority {} message: {:?}", priority, receipt.status);
     }
 
     // Receive all messages
@@ -126,7 +121,10 @@ async fn main() -> Result<()> {
     // ========== Final Stats ==========
     let final_stats = broker.stats();
     println!("\n--- Final Broker Stats ---");
-    println!("  Registered endpoints: {}", final_stats.registered_endpoints);
+    println!(
+        "  Registered endpoints: {}",
+        final_stats.registered_endpoints
+    );
     println!("  Messages delivered: {}", final_stats.messages_delivered);
     println!("  Routes configured: {}", final_stats.routes_configured);
